split tests to seperate files and update yaml
